requirements.txt
PyYAML==3.11
fitbit==0.1.2
google-api-python-client==1.3.1
httplib2==0.9
oauth2client==1.4.3
oauthlib==0.7.2
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-dateutil==2.3
requests==2.5.0
requests-oauthlib==0.4.2
rsa==3.1.4
simplejson==3.6.5
six==1.8.0
uritemplate==0.6
wsgiref==0.1.2
